requirements.txt
numpy==1.14.5
Pillow==5.4.1
